Build WASM file for web and one for non-web environment
The web-sys crate provides access to Web APIs, like Document and
console.  We don't have those APIs in non-web environments. Importing a
WASM file that uses the `web-sys` crate would fail with:

	Traceback (most recent call last):
	File "python/wasm-with-wasmer.py", line 8, in <module>
	    servus = Instance(wasm_bytes)
	RuntimeError: Failed to instantiate the module:
	    link error: Import not found, namespace: ./servus.js, name: __wbindgen_throw

Therefore there are 2 implemententations of the log! macro: 1 that
writes to the browsers console and 1 that does nothing. Depending
on the SERVUS_WEB build variable one of them is build.

Unfortunately this leads to 2 WASM binaries. For now I put the WASM
files targeted for web in pkg/. And the WASM files targeted for non-web
environment are put in python/pkg/.

Ideally I want make this a runtime check so I can create 1 WASM binary.
But in order to do that I need to verify from WASM if the WASM file is
being executed in a browser or not. I am not sure if that is possible.